requirements.txt
openai==0.28
langchain
tiktoken
pinecone-client
streamlit